requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.3
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
decorator==5.1.1
Flask==2.3.2
Flask-Cors==3.0.10
frozenlist==1.3.3
google-api-core==2.17.1
google-auth==2.28.2
google-cloud-aiplatform==1.43.0
google-cloud-bigquery==3.18.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.15.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.1
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.26.4
openai==0.27.6
packaging==24.0
praw==7.7.1
prawcore==2.4.0
proto-plus==1.23.0
protobuf==4.25.3
py==1.11.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
requests==2.30.0
retry==0.9.2
rsa==4.9
shapely==2.0.3
six==1.16.0
tqdm==4.65.0
update-checker==0.18.0
urllib3==2.0.2
websocket-client==1.7.0
Werkzeug==2.3.4
yarl==1.9.2
zipp==3.15.0
